requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
langchain==0.0.350
langchain-openai==0.0.2
langchain-community==0.0.10
openai==1.3.8
requests==2.31.0
pymupdf==1.23.8
pdfplumber==0.10.3
python-multipart==0.0.6
pyyaml==6.0.1
chromadb==0.4.18
faiss-cpu==1.7.4
numpy==1.24.3
python-dotenv==1.0.0
aiofiles==23.2.1
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
